LayoutTests/http/tests/security/javascriptURL/xss-DENIED-from-javascipt-url-in-foriegn-domain-window-open-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/javascriptURL/xss-DENIED-from-javascipt-url-in-foriegn-domain-window-open.html from frame with URL http://localhost:8000/security/javascriptURL/resources/foreign-domain-javascipt-url-accessor-opened-frame.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: line 1: TypeError: Undefined value
Opener Frame

Pass: Cross frame access from a javascript: URL on a different domain was denied.
